Add files via upload
Online Voting System can allow users to cast their ballot from anywhere in the world. The app would work by allowing users to create an account and then log in to their account on the day of the election. Once they have logged in, they would be able to see all of the candidates and issues on the ballot. They would then be able to select their choices and submit their ballot. The app would also allow users to view results and get updates on election night.